Ajout fonction, login, logout, session ... et ajout commentaires

+++ view/forum/TopicsByCategorie.php
@@ -33,11 +33,13 @@ if (empty($topics)) { ?>
                     <a href="index.php?ctrl=forum&action=listPostsByTopic&id=<?= $topic->getId() ?>">
                         <?= $topic->getTitre() ?>
                     </a>
-                    <form action="index.php?ctrl=forum&action=updateTopic&id=<?= $topic->getId() ?>" method="post">
-                        <input type="text" name="titre" placeholder="Modifier le titre">
-                        <input type="submit" name="updateTopic" value="Modifier">
-                    </form>
-                    <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>">Supprimer</a>
+                    <?php if ((App\Session::isAdmin()) || (isset($_SESSION["user"]) && $_SESSION["user"]->getId() == $topic->getUser()->getId())) { ?>
+                        <form action="index.php?ctrl=forum&action=updateTopic&id=<?= $topic->getId() ?>" method="post">
+                            <input type="text" name="titre" placeholder="Modifier le titre">
+                            <input type="submit" name="updateTopic" value="Modifier">
+                        </form>
+                        <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>">Supprimer</a>
+                    <?php } ?>
                 </td>
                 <td>
                     <a href="">
@@ -60,6 +62,6 @@ if (empty($topics)) { ?>
     <form action="index.php?ctrl=forum&action=addTopic&id=<?= $categories->getId() ?>" method="post">
         <input type="text" name="titre" placeholder="Sujet :">
         <textarea name="texte" id="" cols="30" rows="10" placeholder="Message"></textarea>
-        <input type="submit" name="addTopic" value="OK">
+        <input type="submit" name="addTopic" value="Poster">
     </form>
 <?php } ?>
